add docker file and shift some code to exec
